requirements.txt
Flask
requests
psycopg2-binary
python-dotenv
gunicorn # Recommended for production deployment on Render.com